Move basic codec class into codec_base
BaseCodecProject now lives in bin/codec_base.py so it sorts beside
codec_platforms.py and the codec_<family>.py modules instead of hiding
in the gst_utils grab-bag. PipelineInterface stays in gst_utils - the
transform scripts implement it too. Includes the black re-wrap of the
renamed md5 test module's patch strings.
