+++ contrib/checkbox-ce-oem/checkbox-provider-ce-oem/bin/codec_dragonwing.py
@@ -22,9 +22,9 @@ scenarios."""
 
 import argparse
 
+from codec_base import BaseCodecProject
 from gst_utils import (
     GST_LAUNCH_BIN,
-    BaseCodecProject,
     GStreamerEncodePlugins,
     get_test_file_path_by_params,
 )
