+++ contrib/checkbox-ce-oem/checkbox-provider-ce-oem/bin/codec_imx.py
@@ -21,9 +21,9 @@
 
 import argparse
 
+from codec_base import BaseCodecProject
 from gst_utils import (
     GST_LAUNCH_BIN,
-    BaseCodecProject,
     GStreamerEncodePlugins,
     generate_artifact_name,
     get_test_file_path_by_params,
