+++ contrib/checkbox-ce-oem/checkbox-provider-ce-oem/bin/gst_utils.py
@@ -165,63 +165,6 @@ class PipelineInterface(ABC):
         pass
 
 
-class BaseCodecProject(PipelineInterface):
-    """
-    The basic encoder-PSNR project every platform project inherits,
-    like the camera platform classes inherit their base camera class.
-
-    It carries the shared fields and default behaviors; a platform
-    project overrides only what differs:
-      - fill self._pipeline_builders with its codec -> builder mapping
-        (build_pipeline dispatches through it)
-      - override artifact_file / psnr_reference_file when the default
-        does not fit
-    """
-
-    def __init__(
-        self,
-        platform: str,
-        codec: str,
-        width: int,
-        height: int,
-        framerate: int,
-        color_space: str = "",
-        mux: str = "",
-    ) -> None:
-        self._platform = platform
-        self._codec = codec
-        self._width = width
-        self._height = height
-        self._framerate = framerate
-        self._color_space = color_space
-        self._mux = mux
-        self._golden_sample = ""
-        self._artifact_file = ""
-        # codec value -> bound pipeline builder, filled by the platform
-        self._pipeline_builders = {}
-
-    @property
-    def artifact_file(self) -> str:
-        if not self._artifact_file:
-            self._artifact_file = generate_artifact_name()
-        return self._artifact_file
-
-    @property
-    def psnr_reference_file(self) -> str:
-        return self._golden_sample
-
-    def build_pipeline(self) -> str:
-        """
-        Build the GStreamer command via the platform's builder mapping.
-        """
-        builder = self._pipeline_builders.get(self._codec)
-        if builder is None:
-            raise SystemExit(
-                "Error: unknow encoder '{}' be used".format(self._codec)
-            )
-        return builder()
-
-
 def delete_file(file_path: str = "") -> None:
     try:
         if os.path.exists(file_path):
